Create "Smoke" suite
