+++ src/org/labkey/test/categories/Smoke.java
@@ -1,0 +1,8 @@
+package org.labkey.test.categories;
+
+/**
+ * Short suite to verify installed modules are well formed
+ */
+public abstract class Smoke extends Test
+{
+}
